+++ src/js/utils.js
@@ -127,22 +127,33 @@ function readCookie( name ) {
         return cookies[ name ];
     }
 
-    const parts = document.cookie.split( '; ' );
-    cookies = {};
+    // In enketo-validate and perhaps other contexts, enketo-core is used in an empty page in a headless browser
+    // In such a context document.cookie throws an 'Access is denied for this document' error.
+    try {
+        const parts = document.cookie.split( '; ' );
+        cookies = {};
 
-    for ( let i = parts.length - 1; i >= 0; i-- ) {
-        const ck = parts[ i ].split( '=' );
-        // decode URI
-        ck[ 1 ] = decodeURIComponent( ck[ 1 ] );
-        // if cookie is signed (using expressjs/cookie-parser/), extract value
-        if ( ck[ 1 ].substr( 0, 2 ) === 's:' ) {
-            ck[ 1 ] = ck[ 1 ].slice( 2 );
-            ck[ 1 ] = ck[ 1 ].slice( 0, ck[ 1 ].lastIndexOf( '.' ) );
+        for ( let i = parts.length - 1; i >= 0; i-- ) {
+            const ck = parts[ i ].split( '=' );
+            // decode URI
+            ck[ 1 ] = decodeURIComponent( ck[ 1 ] );
+            // if cookie is signed (using expressjs/cookie-parser/), extract value
+            if ( ck[ 1 ].substr( 0, 2 ) === 's:' ) {
+                ck[ 1 ] = ck[ 1 ].slice( 2 );
+                ck[ 1 ] = ck[ 1 ].slice( 0, ck[ 1 ].lastIndexOf( '.' ) );
+            }
+            cookies[ ck[ 0 ] ] = decodeURIComponent( ck[ 1 ] );
         }
-        cookies[ ck[ 0 ] ] = decodeURIComponent( ck[ 1 ] );
+
+        return cookies[ name ];
+
+    } catch( e ){
+        console.error( 'Cookie error', e );
+
+        return null;
     }
 
-    return cookies[ name ];
+
 }
 
 /**