+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+----------------------
+##### Changed
+- Let cookie access attempt fail silently in context where document.cookie is not accessible and throws an exception.
+
 [5.16.4] - 2020-07-23
 ----------------------
 ##### Changed