+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[Unreleased]
+[5.16.5] - 2020-08-03
 ----------------------
 ##### Changed
 - Let cookie access attempt fail silently in context where document.cookie is not accessible and throws an exception.